refactor: convert blog admin to RSC
